Move RuntimeContext to resolver package
