+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/YamlCachedTestBase.java
@@ -3,6 +3,7 @@ package au.csiro.pathling.test.yaml;
 import au.csiro.pathling.fhirpath.context.ResourceResolver;
 import au.csiro.pathling.test.yaml.resolver.CachingResolverBuilder;
 import au.csiro.pathling.test.yaml.resolver.ResolverBuilder;
+import au.csiro.pathling.test.yaml.resolver.RuntimeContext;
 import jakarta.annotation.Nonnull;
 import java.util.Collections;
 import java.util.HashMap;
