+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/YamlTestArgumentProvider.java
@@ -8,11 +8,12 @@ import au.csiro.pathling.test.yaml.YamlTestDefinition.TestCase;
 import au.csiro.pathling.test.yaml.annotations.YamlTest;
 import au.csiro.pathling.test.yaml.annotations.YamlTestConfiguration;
 import au.csiro.pathling.test.yaml.executor.DefaultYamlTestExecutor;
+import au.csiro.pathling.test.yaml.executor.EmptyYamlTestExecutor;
 import au.csiro.pathling.test.yaml.executor.YamlTestExecutor;
 import au.csiro.pathling.test.yaml.resolver.ArbitraryObjectResolverFactory;
 import au.csiro.pathling.test.yaml.resolver.EmptyResolverFactory;
 import au.csiro.pathling.test.yaml.resolver.FhirResolverFactory;
-import au.csiro.pathling.test.yaml.executor.EmptyYamlTestExecutor;
+import au.csiro.pathling.test.yaml.resolver.RuntimeContext;
 import jakarta.annotation.Nonnull;
 import java.io.File;
 import java.util.HashMap;
