+++ fhirpath/src/test/java/au/csiro/pathling/test/yaml/resolver/CachingResolverBuilder.java
@@ -1,7 +1,6 @@
 package au.csiro.pathling.test.yaml.resolver;
 
 import au.csiro.pathling.fhirpath.context.ResourceResolver;
-import au.csiro.pathling.test.yaml.RuntimeContext;
 import jakarta.annotation.Nonnull;
 import java.util.Map;
 import java.util.function.Function;
